% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeSmellsChecker.R
\name{codeSmellsChecker}
\alias{codeSmellsChecker}
\title{Function which scans any code file, and checks it for various code conventions}
\usage{
codeSmellsChecker(codePath, lineLengThresh = 100)
}
\arguments{
\item{codePath}{: path to file to be checked}

\item{lineLengThresh}{: desired soft max for line length
Things checked for
\itemize{
\item un-spaced commas
\item un-spaced operators
\item excessively long lines
}}
}
\description{
Function which scans any code file, and checks it for various code conventions
}
