% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filePullPush.R
\name{filePathedTableRead}
\alias{filePathedTableRead}
\title{Load a dataset from a specified path}
\usage{
filePathedTableRead(filePathIn, fileName, sep1 = ",")
}
\arguments{
\item{filePathIn}{Path to the input file}

\item{fileName}{File name and extension}
}
\value{
dataset
}
\description{
This function loads a file from a specified path, where the file
type is also specified
}
