% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filePullPush.R
\name{filePathedTableWrite}
\alias{filePathedTableWrite}
\title{Save out a dataset to a specified path}
\usage{
filePathedTableWrite(
  dataSet,
  filePathOut,
  fileName,
  extension = ".txt",
  sep1 = ",",
  row.names1 = FALSE,
  col.names1 = TRUE
)
}
\arguments{
\item{dataSet}{the dataset to be written out (as .txt)}

\item{filePathOut}{Path to where the file should be written}

\item{fileName}{File name and extension}
}
\value{
nothing, writes file
}
\description{
This function writes out a file to a specified path, defaulting to .txt
}
