% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filePullPush.R
\name{filePathedToRead}
\alias{filePathedToRead}
\title{Load a dataset from a specified path}
\usage{
filePathedToRead(filePathIn, fileName)
}
\arguments{
\item{filePathIn}{Path to the input file}

\item{fileName}{File name and extension}
}
\value{
filepath to data file
}
\description{
This function returns the path to a file from a specified path, where the file
is also specified
}
