% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NZPoliceColours.R
\docType{data}
\name{policeColours}
\alias{policeColours}
\title{Vector of NZ Police colours
Colours found pg.15 https://tenone.police.govt.nz/sites/default/files/documents/2017-12/NZ\%20Police\%20-\%20Brand\%20Style\%20Guide.pdf
Will require being on Intranet to access that file}
\format{
An object of class \code{character} of length 8.
}
\usage{
policeColours
}
\description{
Vector of NZ Police colours
Colours found pg.15 https://tenone.police.govt.nz/sites/default/files/documents/2017-12/NZ\%20Police\%20-\%20Brand\%20Style\%20Guide.pdf
Will require being on Intranet to access that file
}
\examples{
mycolors["PriBlue"]

}
\keyword{datasets}
